requirements.txt
google-generativeai>=0.1.0
pdfplumber>=0.10.1
chromadb>=0.4.13
python-dotenv>=1.0.0
tqdm>=4.66.1
numpy>=1.24.0
pandas>=2.0.0
ipywidgets>=8.0.0  # For interactive notebook components
matplotlib>=3.7.0  # For visualizations